pyproject.toml
[project]
name = "super_hubo"
readme = "README.md"
requires-python = ">=3.10"
version = "0.1"
dependencies = [
    "qubovert",
    "numpy",
    "planqk-service-sdk"
]
